# jsons/swagger.json
{"openapi":"3.0.0","paths":{"/api/jwt/login":{"post":{"operationId":"AuthController_jwtLogin","summary":"","description":"See test/e2e/local-auth.spec.ts\nneed username, password in body\nskip guard to @Public when using global guard\nSee test/e2e/jwt-auth.spec.ts","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/loginDTO"}}}},"responses":{"201":{"description":""}}}},"/api/jwt/check":{"get":{"operationId":"AuthController_jwtCheck","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}}}},"/api/aws/instance/list":{"get":{"operationId":"InstanceController_list","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}}}},"/api/aws/instance/status":{"post":{"operationId":"InstanceController_status","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}}}},"/api/exchange":{"get":{"operationId":"ExchangeController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"post":{"operationId":"ExchangeController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateExchangeDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/exchange/pairs/{code}":{"get":{"operationId":"ExchangeController_getPairs","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["exchange"],"security":[{"bearer":[]}]}},"/api/exchange/currency_rates":{"get":{"operationId":"ExchangeController_getCurrencyRates","parameters":[],"responses":{"200":{"description":""}},"tags":["exchange"],"security":[{"bearer":[]}]}},"/api/trades":{"post":{"operationId":"TradesController_create","parameters":[],"responses":{"201":{"description":""}},"tags":["trades"],"security":[{"bearer":[]}]}},"/api/trades/history":{"get":{"operationId":"TradesController_getHistory","parameters":[],"responses":{"200":{"description":""}},"tags":["trades"],"security":[{"bearer":[]}]}},"/api/trades/actives":{"get":{"operationId":"TradesController_getActivesTrades","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["trades"],"security":[{"bearer":[]}]}},"/api/trades/{id}/status":{"get":{"operationId":"TradesController_getTradeStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["trades"],"security":[{"bearer":[]}]}},"/api/trades/{id}/add-fund":{"post":{"operationId":"TradesController_addFundTrade","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["trades"],"security":[{"bearer":[]}]}},"/api/trades/{id}/close-by-market":{"get":{"operationId":"TradesController_closeByMarket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["trades"],"security":[{"bearer":[]}]}},"/api/trade-user/me":{"get":{"operationId":"TradeUserController_getMe","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["trade-user"],"security":[{"bearer":[]}]}},"/api/trade-user/accounts":{"get":{"operationId":"TradeUserController_getAccounts","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["trade-user"],"security":[{"bearer":[]}]}},"/api/exchange/{id}":{"get":{"operationId":"ExchangeController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"patch":{"operationId":"ExchangeController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateExchangeDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"delete":{"operationId":"ExchangeController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}}},"info":{"title":"Cats example","description":"The cats API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","in":"header","type":"http"}},"schemas":{"loginDTO":{"type":"object","properties":{"identifier":{"type":"string"},"password":{"type":"string"}},"required":["identifier","password"]},"CreateExchangeDto":{"type":"object","properties":{}},"UpdateExchangeDto":{"type":"object","properties":{}}}}}